refactor initialize; options now supports plugins
