+++ javascript/index.js
@@ -1,6 +1,6 @@
 import packageInfo from '../package.json'
 import { perform, performAsync } from './cable_ready'
-import { initialize } from './elements'
+import { defineElements } from './elements'
 import { shouldMorphCallbacks, didMorphCallbacks } from './morph_callbacks'
 
 import * as Plugins from './plugins'
@@ -19,6 +19,27 @@ import StreamFromElement from './elements/stream_from_element'
 import UpdatesForElement from './elements/updates_for_element'
 import SubscribingElement from './elements/subscribing_element'
 import CableConsumer from './cable_consumer'
+
+const initialize = (initializeOptions = {}) => {
+  const { consumer, onMissingElement, plugins } = initializeOptions
+
+  if (consumer) {
+    CableConsumer.setConsumer(consumer)
+  } else {
+    console.error(
+      'CableReady requires a reference to your Action Cable `consumer` for its helpers to function.\nEnsure that you have imported the `CableReady` package as well as `consumer` from your `channels` folder, then call `CableReady.initialize({ consumer })`.'
+    )
+  }
+
+  if (onMissingElement) MissingElement.set(onMissingElement)
+
+  if (plugins)
+    Object.keys(plugins).forEach(plugin =>
+      Plugins.register(plugin, plugins[plugin])
+    )
+
+  defineElements()
+}
 
 export {
   Utils,
